delegate vertex color edit callbacks to control panel directly
